Simplify dependencies
Preact and @emotion/core are now regular dependencies. This makes the
bundle size a little larger, but it really simplifies things. Sharing
the Preact dep with frontend doesn't seem to work well (I'm not sure but
I think it's a version thing, we saw something similar with automat
work) but using window.guardian.automat to share deps does work, so we
could always revert to that if needs be.