Add Ophan button click tracking behaviour
Platforms provide submitComponentEvent implementation to component as a
prop. This allows us to handle the click handling functionality in here,
which reduces duplication and arguably makes it easier to write tests.

Be robust if passed click handler callbacks throw errors. For example if
one happens to throw an error, we still want to run the other. In future
it’d be good to pass an error handler to the component which could e.g.
log to Sentry if this kind of thing happens.

Also increase the test coverage around the component a little.