Make @guardian/types a direct dependency
Not a dev dependency. I ran into an issue where the version specified by
DCR was ^0.4.4 but @guardian/braze-components wants ^0.4.5. I’d thought
that would resolve to 0.4.5 in DCR but I think yarn doesn’t take into
account the version required by braze-components because it’s a dev
dependency. We ship TypeScript types with this package, which refer to
types in @guardian/types, so I think a direct dependency makes sense. I
think a peer dependency would also work, but this feels simpler so
trying this approach first.