Add pan-domain auth and enable image picker
This adds client side auth using the pan-domain cookie, to reduce
friction when interacting with the Grid image picker. You need to be
logged in for the Grid picker to work. We check your logged in status by
pinging the login service and including the pan domain cookie in the
request. This is all implemented as a Storybook plugin.

We now also call the image signing service before presenting the final
image URL back to the user in Storybook as a component prop/knob.

Co-authored-by: Michael Clapham <michaeljclapham@gmail.com>