Add a second entry point to the package
This will allow us to import @guardian/braze-components/logic and _only_
get the functionality downstream of that entry point, rather than the
entire package.

Following a conversation in Client Side I discovered we need to include
a nested package.json for the logic code to allow requesters to use the
correct file (esm vs cjs).