Reflect the prop validation we do on platform
It is unfortunate that here we're duplicating the checks that we do
on-platform before rendering the Braze epic. This should be addressed
properly, but in the meantime I'm keen to have Storybook reflect the
platform behaviour so this doesn't cause confusion for marketing. This
isn’t an issue with the banner components as the component itself does
these checks. Here we’re using the support-dotcom-components epic and
aren’t in control of the prop validation. Possibly an argument for
having our own epic component.