BrazeMessages returns first valid message in cache (#155)
BrazeMessages returns first valid message in cache.

Before PR
All messages from Braze are added to a local storage backed queue. When rendering a component for a slot (Body end or Banner), the first message from the queue is used to try to render a component. If that message is invalid no component is rendered for the slot.

After PR
When searching the queue for a braze message to pass to the component to render, filtering occurs to remove messages that are not valid according to that components canRender function.

Invalid messages are still added to the queue, they are filtered out when accessing the queue. This means in the future if a message is published that currently counts as invalid, but is intended to be shown, we could publish a hotfix that makes it valid, and users who had previously cached it, would now see a component based on that message.

Can Render Tree shaking
The can render functions for each component and for all the components are in separate canRender files. This is to enable rollup to treeshake, so that the logic package can load the canRender methods to detect if a component can render, without having to load the entire component. This keeps the size of the main frontend / DCR bundle smaller.

Co-authored-by: Tom Wey <tom.wey@theguardian.com>
Co-authored-by: Tom Wey <tjmw@users.noreply.github.com>