Group stories by slot
We support two slots currently - Banner and EndOfArticle. It makes sense
to group components in Storybook by these slots. Also setting the story
name to match the component name seems to have the effect of removing an
extra layer from the hierarchy which I think makes it simpler to
navigate.