Add first prop to component
The header is now configurable. I updated the story to provide a knob to
control the header and changed a couple of things along the way. We’re
now using the top level BrazeMessageComponent prop to render the news-
letter epic. I’m not sure sharing one “factory” component like this
across slots makes sense, but it gets us started.