US Newsletter Epic component (#175)
* Start of US Newsletter Epic component

* Style changes

* Add first prop to component

The header is now configurable. I updated the story to provide a knob to
control the header and changed a couple of things along the way. We’re
now using the top level BrazeMessageComponent prop to render the news-
letter epic. I’m not sure sharing one “factory” component like this
across slots makes sense, but it gets us started.

* Update this to fit with new component story hierarchy

* Make frequency text a prop

* Add paragraphs as separate props

Keep it simple and only add 2 for now as that’s all the designs include.
Only paragraph1 is required.

* Update clock icon with version from src-icons

Not actually available yet, but I’ve opened a PR to add a new SvgClock
icon based on the markup in the source figma:
https://github.com/guardian/source/pull/843

* Fix sign up button style

* Change newsletter epic layout on mobile

* Fix lint issues

* Newsletter epic padding

* Remove button hover colour

* Work in progress storybook folder

Co-authored-by: Tom Wey <tom.wey@theguardian.com>