Add support for subscribeToNewsletterProp to epics
This PR adds support for a subscribeToNewsletter prop to the newsletter
epic components. This prop has the following type:

export type NewsletterSubscribeCallback = (id: string) => Promise<void>;

The implementation for this prop will be supplied by the platforms. The
function is invoked when a user clicks on the "Sign up" button.

Note that the in-progress state and failures states aren't complete yet
as we're waiting on design input.

Note that the subscribe function is a prop of BrazeMessageComponent.
Currently both epics and banners go through this factory component
(which is why I’ve added this new prop to all the component stories) but
I think we should split BrazeMessageComponent into two separate
factories - one for banners and one for epics.

Co-authored-by: Tom Wey <tom.wey@theguardian.com>