Pass subscribeToNewsletter as a callback
This will be provided by DCR. The interface is that it takes a
newsletter id and returns a promise that resolves when the subscribe
is successful.

The callback is a prop of BrazeMessageComponent. Currently both epics
and banners go through this factory component (which is why I’ve added
this new prop to all the component stories) but I think we should split
BrazeMessageComponent into two separate factories - one for banners and
one for epics.