Update newsletter epic `canRender` functions (#199)
* Make ophanComponentId required by epic canRenders

Make the ophanComponentId a required prop in all newsletter epics. It is
already required by the regular contributions epic. We use it on the
platform for tracking VIEWs and INSERTs and if it’s missing we should
not render the component. In a few cases I also moved it from the top
level Props to the BrazeMessageProps since it comes from Braze as a key
value pair.

* Validate the NewsletterEpic imageUrl

In the same way we do for the AppBanner.