Split banner and epic factories (#194)
* Split the BrazeMessageComponent factory in two

Now we have two factory components, one per slot. This has the advantage
of having to only pass in the set of props that we might need for the
components for a given slot. So we now no longer need to pass subscribe
ToNewsletter to the banner component factory, as none of the underlying
components require it.

I also removed some unused props from the common end of article props
interface which weren’t used by any of the underlying components.

* We need to export these to be used by the platform

Co-authored-by: Michael Clapham <michael.clapham@guardian.co.uk>