Support HTML in paragraphs
We have use cases where there are links in the paragraph text. Before
this change any HTML in the paragraph was escaped. This change makes us
behave the same as the RRCP component which we’re switching away from.

Also add a link to the default props in storybook so that if this breaks
in future Chromatic will let us know.