Fix invalid placeholder detection logic
Also added some tests as documentation. Previously the code to determine
whether there were any invalid placeholders was always returning false,
this fixes the check so that invalid placeholders are correctly flagged.

The fields we support placeholder replacement in are the paragraphs and
highlightedText. Update the call to containsNonAllowedPlaceholder in
the canRender to reflect this.