Create separate EpicWithHeaderImage component
We’re treating our first usage of the epic with an image as an
experiment, so short term we’ll hardcode the image avoiding the overhead
of adding proper image support (via the picker). If the test goes well
we can make headerImageUrl a properly supported part of the Epic
component.

In this iteration headerImageUrl is not passed to the parent Epic
component as part of the brazeMessageProps, so there’s no way to set it
from the Braze extras data (i.e. it’s private).