Create new EpicWithSpecialHeader component
We’re treating our first usage of the epic with this additional header
section as an experiment, so short term we’ll hardcode the image
avoiding the overhead of adding proper image support (via the picker),
and text. If the test goes well we can make these properties properly
supported in the Epic component.