Use master crop image source for transformations
From talking to other developers who understand the image transformation
service well, we should always be using the master crop as the source
for image transformations.

I've also added in the quality and auto=format params. Auto format does
causes the image transformation service to do the right thing with
regards to converting to either webp or png8.

My goal when looking at the transformations we were applying was to see
if I could get the image size down any further, which I wasn't
successful in doing. Of the three images two a very marginally smaller
and one is very marginally larger than what we had before.