Fix some uses of any in LocalMessageCache
Fixing this properly caused a bit of a ripple effect, but I think it
leaves us in a better state, and our types reflect how we use Braze more
accurately. We always use HtmlMessage when configuring in-app messages,
this assumption already existed in the code, but now the types reflect
this too.

As part of this work I removed the in-memory cache, since this is no
longer used.