Refactor to make differences from common styles clearer
We were previously duplicating styles from commonStyles and tweaking
them slightly which makes it harder to see what’s actually different.
Instead, I’ve extracted the parts that are different and am using them
to override the common styles. I’d like to get to the bottom of _why_
these differences exist but that might be for a future PR.

Chromatic passes at this stage, which gives confidence that this hasn't
had any visual impact.