Style changes to support Guardian in 2021 banner
Colour changes (orange background, blue link button). Also some
simplifications to support the new colours (e.g. removing hover effects
in a couple of places).

I also updated default image for BannerWithLink. Just to make it easier
to preview the new colour palette with the image to be used. This is the
image provided, but it's probably not high enough resolution. I need to
request a higher res one.