Tweaks to BannerWithLink image styles
This has the effect of increasing the sizing of the image, so there’s a
bit less space all around. The max height was kicking in previously,
reducing the max possible height of the image, so I’ve increased this
slightly. Also removed the negative top margin with a positive value to
ensure we’ve always got a little bit of space at the top. I think this
gives the best appearance, balancing different image sizes (ratios) and
different text content lengths.

Also updated the default image in the BannerWithLink story and removed a
duplicated CSS property.