Add (optional) remind me button to Epic component
The remind me button will appear if button text, confirmation text and
confirmation header text are provided as props. No tracking yet, that
will be added in a subsequent commit. Once the tracking is implemented,
the Braze canvas will know if a reader clicked remind me, so we'll be
able to schedule the reminder email.

Note the (breaking) change that we now need the consumer to provide
@guardian/source-react-components-development-kitchen. We get the
divider Lines component from there, used in the confirmation message.