Click tracking on regular epic (including remind me button)
This includes a refactoring to expose a single trackClick callback to
the individual epic components, instead of submitComponentEvent and
logButtonClickWithBraze separately. This keeps the underlying code
simpler and also DRYs things up a bit. I've only done this refactoring
for epics here. Banners will be done in a separate PR.