Introduce BrazeCards, backed by Braze's Content Cards
BrazeCards is a mechanism for providing persistent 'notification'
messages, in contrast to the existing single-impression (In-App
Message-backed) BrazeMessages behaviour.

This PR introduces the basic abstraction for content cards, without
providing any specific slot component. BrazeCards looks very similar
to BrazeMessages, but exposes a different interface.

- it returns collections of cards instead of selecting a single message
- Braze's SDK serves cards out its own cache, so there's no `promise`
- the above means we can also skip our own cache implementation

This change also splits the representation of 'slots' into
`MessageSlot`s and `CardSlot`s. This change encodes in the types the
distinction between slots which will display the single
highest-priority message (if available), versus those that are able to
prompt the user on multiple counts simultaneously.

I expect in the future we'd consider setting up notification slots to
support multiple message streams of differeing priority, but this is
something we should add as we develop a better intuition of how this
fature will be used. Well-considered [prior
art](https://developer.android.com/training/notify-user/channels)
exists for these features, when the need arises.