Content cards spike (#316)
* Add testwatch command to project scripts

* Introduce BrazeCards, backed by Braze's Content Cards

BrazeCards is a mechanism for providing persistent 'notification'
messages, in contrast to the existing single-impression (In-App
Message-backed) BrazeMessages behaviour.

This PR introduces the basic abstraction for content cards, without
providing any specific slot component. BrazeCards looks very similar
to BrazeMessages, but exposes a different interface.

- it returns collections of cards instead of selecting a single message
- Braze's SDK serves cards out its own cache, so there's no `promise`
- the above means we can also skip our own cache implementation

This change also splits the representation of 'slots' into
`MessageSlot`s and `CardSlot`s. This change encodes in the types the
distinction between slots which will display the single
highest-priority message (if available), versus those that are able to
prompt the user on multiple counts simultaneously.

I expect in the future we'd consider setting up notification slots to
support multiple message streams of differeing priority, but this is
something we should add as we develop a better intuition of how this
fature will be used. Well-considered [prior
art](https://developer.android.com/training/notify-user/channels)
exists for these features, when the need arises.

* Abstract BrazeCards to allow a null implementation

Also corrects the entrypoint's imports.

* Braze SDK needs to be a type-only import for server-side

The import was missing the `type` keyword, which is essential to
prevent browser-specific code from being accidentally run on the
server.

* Swap `ControlCard` for `Card` in Braze card handling

ControlCard instances can't be created from the UI for campaigns
directly, they really are just the 'control'. Since the base Card
class contains everything we need we don't actually need to specify
any particular subclass, which makes this much more flexible.

* Address PR comments

Also removes the (redundant) `testwatch` yarn script.