Add support for multiple section filters
We've supported specifying a section key/value pair in the message
configuration in Braze for a while. Up until now there was a limitation
that you could only provide a single section, as the comparison with the
page section was a simple equality check.

This adds support for specifying multiple sections separated by the |
(pipe) character. For example, if you wanted your message to appear on
environment OR science articles you would supply a key/value pair like
this:

"section" -> "environment|science"