Source package upgrades (#369)
DCR is upgrading @guardian/source-* libs. We rely on the platform providing
source libs as peerDependencies. Bump the versions in here so we match what DCR
is providing (there's an open question of what to do for frontend, which we are
also consumed by, and which hasn't upgraded source libs yet).

There were a couple of styling tweaks for underlines on the secondary button in
AppBanner and for the info icon. I also fixed some deprecated exports (SvgInfo
-> SvgInfoRound and usage of background.