Update peer deps to reflect what's on the platforms (#387)
Update peer dependencies and dev dependencies to the latest major versions.
This will require a major version bump. We're ignoring the recommendation to
keep peer deps as broad as possible since the only consumers are codebases we
can control. The versions of our peer deps will need to be updated on the
platforms at the same time, to reflect the change

Also fixed some other eslint peer dependency warnings by upgrading some other
eslint packages. Require a change to the eslint config to not complain about
the css emotion prop.