Upgrade np
When we tried to publish to NPM using yarn release we got an error
about the ls-collaborators npm command not existing. It looks like this
command has been renamed in more recent NPM versions. This was the first
publish we've attempted since upgrading Node so I think that's the
culprit. The solution is to upgrade np.