Publish to NPM from CI (#436)
This adds a new step to the main GHA workflow to run yarn changeset publish.

When a PR is merged which is part of a changeset, a new PR will be
automatically opened to create the new release. When this PR is merged to main,
the new release will be published to NPM. If the release PR is not merged and
subsequent PRs are merged with a changeset specified, the release PR will be
updated.