Avoid casting user provided styles for PrimaryCtaButton
Instead accept that they might be missing and let the type guard used by
getColors handle this. The return value from getColors is guaranteed to
have legit ColorValueHex values.

I've decoupled the getColors logic (combine user provided styles with
defaults, in case any user provided styles are missing or invalid) and
the logic to take the colours and convert them to styles.

I've focused on the PrimaryCtaButton here, but the pattern could be
applied to all components which support user provided styles.