Upgrade rollup-plugin-filesize
Even though this is a major version bump, it seems to work in a
backwards compatible way - the filesize reporting when running `yarn
build` looks the same.

Doing this because it removes one of the paths causing us to pull in the
ip package, which currently has an unpatchable vulnerability.