Upgrade Storybook to v8 (#501)
Upgrade Storybook to v8. This major version bump removes the other path which
was bringing in the vulnerable ip package (the first use was removed in #500).

I ran:
$ npx storybook@latest upgrade

And fixed a few remaining issues following:
https://github.com/storybookjs/storybook/blob/next/MIGRATION.md#from-version-7x-to-800.

There were some issues with the Grid add on which needed to be resolved through
trial and error.