Broaden @guardian/source peer dep requirements up to v6 (#506)
Broaden the @guardian/source peer dep range from `^1.0.1` to `>= 1.0.1 < 7`. I've
tested with v6 and the component rendering looks good (I've made an assumption
that things are fine in the intermediate major versions too).

As per our recommendations I've bumped the dev version to the latest acceptable.

This ought to make upgrading incrementally on DCR more straightforward.